backend/crawler/requirements.txt
selenium>=4.15.0
python-dotenv
pika==1.3.2
webdriver-manager==4.0.1
gender-guesser==0.4.0
